requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
pandas>=1.5.0
pyyaml>=6.0
streamlit>=1.28.0
transformers>=4.30.0
xgboost>=1.7.0
lightgbm>=3.3.0
shap>=0.42.0
